メッセージ機能のフロントエンド側機能は実装できました
