+++ src/Chat/ChatPane.css
@@ -1,0 +1,5 @@
+.ChatPane{
+    border:solid 1px black;
+    margin:1em;
+    height:100%;
+}
