+++ src/Chat/ChatPane.tsx
@@ -1,0 +1,31 @@
+import React from "react";
+import ChatMessageComponent from "./ChatMessageComponent";
+import "./ChatPane.css";
+import DefaultIcon from "./../image/userimg/defaultimg.png"
+
+
+const fetchMemberList = async (setMemberList) =>{
+    const inputdata = await fetch(`http://localhost/backend/member`, {
+        method: 'GET',
+        mode: 'cors'
+    })
+    .then(res => res.json())
+    .then(members => {
+        setMemberList(members);
+    });
+}
+
+function Chat(){
+
+    return(
+        <div className="ChatPane">
+            <ChatMessageComponent chatMessage="hogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehoge"/>
+            <ChatMessageComponent chatMessage="hogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehoge"/>
+            <ChatMessageComponent chatMessage="hogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehoge"/>
+            <ChatMessageComponent chatMessage="hogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehoge"/>
+            <ChatMessageComponent chatMessage="hogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehogehoge"/>
+        </div>
+    );
+}
+
+export default Chat;
